Reorganize some parser extensions
